feat: + update quiz validation schemas cy-229
