+++ packages/shared/src/modules/quiz/libs/enums/quiz-answers-rules.enum.ts
@@ -1,6 +1,6 @@
 const QuizAnswersRule = {
 	ANSWERS_LIST_MIN_LENGTH: 1,
-	QUESTION_TEXT_MIN_LENGTH: 3,
+	QUESTION_TEXT_MIN_LENGTH: 2,
 } as const;
 
 export { QuizAnswersRule };
