+++ packages/shared/src/modules/quiz/libs/utilities/utilities.ts
@@ -1,0 +1,1 @@
+export { sanitizeTextInput } from "shared";
