+++ packages/shared/src/modules/quiz/libs/validation-schemas/quiz-submit-answers.validation-schema.ts
@@ -1,4 +1,3 @@
-import { sanitizeTextInput } from "shared";
 import z from "zod";
 
 import {
@@ -6,42 +5,24 @@ import {
 	QuizCategory,
 	QuizValidationMessage,
 } from "../enums/enums.js";
+import { quizAnswerSchema } from "./quiz-shared.validation-schema.js";
 
 const quizAnswersSchema = z.object({
 	answers: z
 		.array(
-			z
-				.object({
-					isSkipped: z.boolean(),
-					questionId: z.number().int().positive(),
-					questionText: z
-						.string()
-						.trim()
-						.min(QuizAnswersRule.QUESTION_TEXT_MIN_LENGTH, {
-							message: QuizValidationMessage.QUESTION_TEXT_MIN_LENGTH,
-						}),
-					selectedOptions: z.array(
-						z.union([z.number().int().positive(), z.string().trim()]),
-					),
-					userInput: z.preprocess(
-						(value) =>
-							typeof value === "string" ? sanitizeTextInput(value) : "",
-						z.string().default(""),
-					),
-				})
-				.refine((answer) => {
-					const { isSkipped, selectedOptions, userInput } = answer;
+			quizAnswerSchema.refine((answer) => {
+				const { isSkipped, selectedOptions, userInput } = answer;
 
-					if (isSkipped) {
-						return true;
-					}
+				if (isSkipped) {
+					return true;
+				}
 
-					const hasAnySelectedOptions =
-						selectedOptions.length >= QuizAnswersRule.ANSWERS_LIST_MIN_LENGTH;
-					const hasUserInput = Boolean(userInput);
+				const hasAnySelectedOptions =
+					selectedOptions.length >= QuizAnswersRule.ANSWERS_LIST_MIN_LENGTH;
+				const hasUserInput = Boolean(userInput);
 
-					return hasAnySelectedOptions || hasUserInput;
-				}, QuizValidationMessage.NON_SKIPPED_QUESTION),
+				return hasAnySelectedOptions || hasUserInput;
+			}, QuizValidationMessage.NON_SKIPPED_QUESTION),
 		)
 		.refine(
 			(answers) =>
